refactor(leaderboard): extract shared page shell and scope hook
Extract duplicated logic from 3 leaderboard pages into reusable abstractions:

1. useLeaderboardScope hook (use-leaderboard-scope.ts):
   - Auth-gated scope initialization
   - Organization/team fetching with orgsLoaded guard
   - Scope validation (reset invalid stored scope)
   - localStorage persistence

2. LeaderboardPageShell component (leaderboard-page-shell.tsx):
   - Loading skeleton
   - Error state
   - Empty state
   - Load-more button
   - Profile dialog wiring

Simplifies each page to only contain its specific selector + useLeaderboard call.

Closes #61
